+++ map_model/lib/map_model.dart
@@ -2,3 +2,4 @@ library map_model;
 
 export 'annotation/model.dart';
 export 'src/map_model.dart';
+export 'src/convert.dart';