+++ map_model/lib/src/map_model.dart
@@ -3,21 +3,23 @@ import 'package:meta/meta.dart';
 abstract class MapModel {
   @protected
   // ignore: non_constant_identifier_names
-  late Map<String, dynamic> $_data;
+  late Map<String, dynamic> $data;
 
-  void useDefault() {}
+  Map<String, Type> $types() => {};
 
-  void validate() {}
+  void $default() {}
 
-  Map<String, dynamic> export() => {};
+  void $validate() {}
 
-  void bindTo(MapModel other) {
-    other.$_data = $_data;
-    validate();
+  Map<String, dynamic> $export() => {};
+
+  void $bindTo(MapModel other) {
+    other.$data = $data;
+    $validate();
   }
 
-  MapModel([Map<String, dynamic>? data]): $_data = data ?? {} {
-    useDefault();
-    validate();
+  MapModel([Map<String, dynamic>? data]) : $data = data ?? {} {
+    $default();
+    $validate();
   }
 }