+++ CHANGELOG.md
@@ -4,3 +4,7 @@
 
 ## 1.0.7
 - No longer providing convert, but verifying the type.
+
+## 1.0.8
+- Add default type convert.
+- Add custom type convert.