+++ map_model/CHANGELOG.md
@@ -17,3 +17,7 @@
 
 ## 1.0.7
 - No longer providing convert, but verifying the type.
+
+## 1.0.8
+- Add default type convert.
+- Add custom type convert.